requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-crispy-forms
crispy-bootstrap5
drf-yasg